Function App/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.26.0
playwright==1.55.0
